Move jib builder code to its own package
